+++ headers/private/system/syscalls.h
@@ -299,7 +299,7 @@ extern ssize_t		_kern_write(int fd, off_t pos, const void *buffer,
 						size_t bufferSize);
 extern ssize_t		_kern_writev(int fd, off_t pos, const struct iovec *vecs,
 						size_t count);
-extern status_t		_kern_ioctl(int fd, ulong cmd, void *data, size_t length);
+extern status_t		_kern_ioctl(int fd, uint32 cmd, void *data, size_t length);
 extern ssize_t		_kern_read_dir(int fd, struct dirent *buffer,
 						size_t bufferSize, uint32 maxCount);
 extern status_t		_kern_rewind_dir(int fd);
